backend/requirements.txt
fastapi
motor
uvicorn
python-jose
